requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.5.0
matplotlib==3.8.4
shap==0.46.0
seaborn==0.13.2
pyyaml==6.0.2
imbalanced-learn==0.12.3
pytest==8.2.2
black==24.8.0
ruff==0.6.9

joblib==1.4.2
fastapi==0.115.0
uvicorn==0.30.6
